+++ changelog.md
@@ -3,5 +3,5 @@
 # 0.0
 
 **03/11/2022**
-> start.bat updated to build
-> created files wz-login/wz-lobby
+1. start.bat updated to build
+2. created files wz-login/wz-lobby