+++ changelog.md
@@ -5,3 +5,4 @@
 **03/11/2022**
 1. start.bat updated to build
 2. created files wz-login/wz-lobby
+3. DB updated