+++ changelog.md
@@ -6,3 +6,6 @@
 1. start.bat updated to build
 2. created files wz-login/wz-lobby
 3. DB updated
+
+**04/11/2022**
+4. Fixed loadingscreen