+++ changelog.md
@@ -10,3 +10,6 @@
 **04/11/2022**
 4. Fixed loadingscreen
 5. Added some features to login
+
+**05/11/2022**
+6. New features to login